Tally resources in a timestamped file.
Resources are now tallied in resources.YYYY-MM-DDTHHMMSS (ISO 8601 format local time).
For backward compatibility and ease of human interaction, the plainly-named "resources" file is now a symbolic link that points to the latest resources file.

This alleviates the difficulty associated with restarts as explained in issue #5.

x/report has been updated to reflect the change in naming scheme.