Ensure some batch scheduler is used.
This fix partially addresses the race condition in issue #12.

What happened is that no METAQ_JOB_RUNNING function was defined.
Bash tried to execute that function, failed, and returned the null string "", which is what a properly-working METAQ_JOB_RUNNING is supposed to do (when it is working properly) to indicate that the job doesn't exist (ie. can be reset).