Add METAQ_SKIP_ON_STOLEN
We found that when running a lot of jobs, a lot of time was spent looking for work, even with `METAQ_SORT=shuf` enabled.
The issue is that the list of things that are considered for a particular folder is only refreshed at the beginning of considering that folder.

We came up with three options:
    -   `false`: don't care, just keep looking in here.
    -   `directory`: some other job is looking in here, so skip this directory.
    -   `reset`: go back to the highest-priority directory and look in there.  If you're already there, at least that causes the list of possible work to refresh.

There may be a more subtle option, which would be to keep looking in this directory, but just refresh the list.  I'm not sure that's simple to implement, I'd have to think further.