METAQ_SLEEP_AFTER_LAUNCH
When a task launches successfully, sleep.

We are encountering an unusual problem on Sierra when multiple MQ jobs run simultaneously.

We hope that spacing out the task launching can solve this problem.

We didn't want to delay for every task, including those that failed to launch, because that could slow things down very dramatically.